\name{pythagoras-package}
\alias{pythagoras-package}
\alias{pythagoras}
\docType{package}
\title{
\packageTitle{pythagoras}
}
\description{
\packageDescription{pythagoras}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pythagoras}
\packageIndices{pythagoras}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{pythagoras}

Maintainer: \packageMaintainer{pythagoras}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
